Allow xhr "retry"
When working with interceptors for authentication (http 401 -> re-authentication -> retry), on the "retry" call, the '__setXHR_' header function has no wrapper anymore. In this case, there is no 'xhr' argument since the function is called directly by angular. Therefore, that function should not be executed (which is not a problem since the events listeners have already been added).